# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nearby Locations API",
    "version": "1.0.0",
    "description": "Find locations whose service radius reaches a given point."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {}
  },
  "paths": {
    "/openapi.json": {
      "get": {
        "summary": "OpenAPI document",
        "tags": [
          "system"
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "system"
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/auth/token": {
      "post": {
        "summary": "Issue a JWT for the given credentials",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "properties": {
                  "username": {
                    "minLength": 1,
                    "type": "string"
                  },
                  "password": {
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "token_type": {
                      "type": "string",
                      "enum": [
                        "Bearer"
                      ]
                    },
                    "expires_in": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "access_token",
                    "token_type",
                    "expires_in"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/locations/search": {
      "get": {
        "summary": "Find locations whose service radius reaches the given point",
        "tags": [
          "locations"
        ],
        "parameters": [
          {
            "schema": {
              "minimum": 0,
              "maximum": 9007199254740991,
              "type": "integer"
            },
            "in": "query",
            "name": "x",
            "required": true
          },
          {
            "schema": {
              "minimum": 0,
              "maximum": 9007199254740991,
              "type": "integer"
            },
            "in": "query",
            "name": "y",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user-location": {
                      "pattern": "^x=(0|[1-9][0-9]*),y=(0|[1-9][0-9]*)$",
                      "type": "string"
                    },
                    "locations": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "coordinates": {
                            "pattern": "^x=(0|[1-9][0-9]*),y=(0|[1-9][0-9]*)$",
                            "type": "string"
                          },
                          "distance": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "coordinates",
                          "distance"
                        ]
                      }
                    }
                  },
                  "required": [
                    "user-location",
                    "locations"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/locations/{id}": {
      "get": {
        "summary": "Get a location by id",
        "tags": [
          "locations"
        ],
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "opening-hours": {
                      "type": "string"
                    },
                    "image": {
                      "type": "string"
                    },
                    "coordinates": {
                      "pattern": "^x=(0|[1-9][0-9]*),y=(0|[1-9][0-9]*)$",
                      "type": "string"
                    },
                    "radius": {
                      "minimum": 1,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "name",
                    "type",
                    "id",
                    "opening-hours",
                    "image",
                    "coordinates",
                    "radius"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ]
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Create or replace a location",
        "tags": [
          "locations"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "properties": {
                  "name": {
                    "minLength": 1,
                    "type": "string"
                  },
                  "type": {
                    "minLength": 1,
                    "type": "string"
                  },
                  "id": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "opening-hours": {
                    "minLength": 1,
                    "type": "string"
                  },
                  "image": {
                    "format": "uri",
                    "minLength": 1,
                    "type": "string"
                  },
                  "coordinates": {
                    "pattern": "^x=(0|[1-9][0-9]*),y=(0|[1-9][0-9]*)$",
                    "type": "string"
                  },
                  "radius": {
                    "minimum": 1,
                    "type": "integer"
                  }
                },
                "required": [
                  "name",
                  "type",
                  "id",
                  "opening-hours",
                  "image",
                  "coordinates",
                  "radius"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "opening-hours": {
                      "type": "string"
                    },
                    "image": {
                      "type": "string"
                    },
                    "coordinates": {
                      "pattern": "^x=(0|[1-9][0-9]*),y=(0|[1-9][0-9]*)$",
                      "type": "string"
                    },
                    "radius": {
                      "minimum": 1,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "name",
                    "type",
                    "id",
                    "opening-hours",
                    "image",
                    "coordinates",
                    "radius"
                  ]
                }
              }
            }
          },
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "opening-hours": {
                      "type": "string"
                    },
                    "image": {
                      "type": "string"
                    },
                    "coordinates": {
                      "pattern": "^x=(0|[1-9][0-9]*),y=(0|[1-9][0-9]*)$",
                      "type": "string"
                    },
                    "radius": {
                      "minimum": 1,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "name",
                    "type",
                    "id",
                    "opening-hours",
                    "image",
                    "coordinates",
                    "radius"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "locations",
      "description": "Location search, detail and upsert"
    },
    {
      "name": "auth",
      "description": "Token issuance"
    },
    {
      "name": "system",
      "description": "Health and meta"
    }
  ]
}